Add JWT Authentication and update blog posts and categories
The JWT authentication process was implemented for the .NET Core project to secure the application's API endpoints. Two new blog posts were created and added to the system. Additionally, the existing blog post was removed, and the list of blog post categories was updated. Lastly, the title on the homepage was changed to "DotNET Evangelist".